% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multidplyr_helpers.R
\name{batch_apply}
\alias{batch_apply}
\title{Helper function to apply a function batchwise to a data frame}
\usage{
batch_apply(df, batch_size = 10000, function_to_apply_to_party_df,
  values_to_assign, packages)
}
\arguments{
\item{df}{data frame}

\item{function_to_apply_to_party_df}{function to apply to party df}

\item{values_to_assign}{values to assign}

\item{packages}{packages to assign}
}
\value{
data frame after applying function
}
\description{
Helper function to apply a function batchwise to a data frame
}
