% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kaggle_helpers.R
\name{construct_submission_name}
\alias{construct_submission_name}
\title{Helper function to construct convenient submission name.}
\usage{
construct_submission_name(model_slug = "no_slug")
}
\arguments{
\item{model_slug}{model slug to paste after timestamp}
}
\value{
convenient submission file name
}
\description{
Helper function to construct convenient submission name.
}
\examples{
construct_submission_name()
construct_submission_name("my_model")

}
