% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detatch_all_packages.R
\name{detach_all_packages}
\alias{detach_all_packages}
\title{Detatch all packages}
\usage{
detach_all_packages()
}
\description{
\code{detach_all_packages} detaches all currently loaded R packages except the base packages.
}
\examples{

\dontrun{
detach_all_packages()
}

}
