% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detatch_all_packages.R
\name{detatch_all_packages}
\alias{detatch_all_packages}
\title{Detatch all packages}
\usage{
detatch_all_packages()
}
\description{
\code{detatch_all_packages} detatches all currently loaded R packages except
the base packages.
}
\examples{

\dontrun{
detatch_all_packages()
}

}
