% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_maxxxy_packages.R
\name{get_maxxxy_packages}
\alias{get_maxxxy_packages}
\title{Get maxxxy packages}
\usage{
get_maxxxy_packages()
}
\description{
\code{get_maxxxy_packages} returns a list of the most used maxxxy packages
}
\examples{
get_maxxxy_packages()

}
