% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_grid.R
\name{multiplot}
\alias{multiplot}
\title{Create multi-panel plots}
\usage{
multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects}

\item{plotlist}{A list of ggplot objects}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored.}
}
\value{
a multi-panel plot

If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
then plot 1 will go in the upper left, 2 will go in the upper right, and
3 will go all the way across the bottom.
}
\description{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
}
\examples{
library(tidyverse)
library(maxxxy)

p1 <- iris \%>\%
  as_tibble() \%>\%
  ggplot(aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point()

p2 <- iris \%>\%
  as_tibble() \%>\%
  ggplot(aes(Sepal.Width, fill = Species)) +
  geom_density(alpha = 0.5)

p3 <- iris \%>\%
  as_tibble() \%>\%
  ggplot(aes(Species, Sepal.Width, fill = Species)) +
  geom_boxplot()

multiplot(p1, p2, p3)

}
